Move CachedExtensionIds and SPI helpers to pg_extension_base
pg_extension_base is the foundational extension in the dependency chain,
making it the natural home for shared infrastructure used by all extensions.

This eliminates code duplication and allows pg_extension_base itself to
use proper SPI helpers with extension owner context switching.

Specific changes:
- Move extension_ids.h/c (CachedExtensionIds infrastructure) from pg_lake_engine
- Move spi_helpers.h from pg_lake_engine
- Add PgExtensionBase cache in separate pg_extension_base_ids.h/c files
- Update SPI_START_EXTENSION_OWNER macro to work with any CachedExtensionIds
- Use SPI_START_EXTENSION_OWNER(PgExtensionBase) in base_worker_launcher.c
  for register/deregister functions to support non-superuser callers
- Update all extension Makefiles to include pg_extension_base headers
