+++ pg_extension_base/include/pg_extension_base/extension_ids.h
