+++ pg_lake_engine/include/pg_lake/extensions/btree_gist.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 
 #define BTREE_GIST "btree_gist"
