+++ pg_lake_engine/include/pg_lake/extensions/pg_lake_table.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 
 #define PG_LAKE_TABLE "pg_lake_table"
 #define PG_LAKE_TABLE_SCHEMA "lake_table"
