+++ pg_lake_engine/src/cleanup/in_progress_files.c
@@ -38,7 +38,7 @@
 #include "pg_lake/locks/advisory_locks.h"
 #include "pg_lake/pgduck/remote_storage.h"
 #include "pg_lake/util/array_utils.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 #include "pg_lake/util/plan_cache.h"
 #include "pg_lake/util/string_utils.h"
 #include "datatype/timestamp.h"
