+++ pg_lake_engine/src/extensions/pg_lake_engine.c
@@ -19,7 +19,7 @@
 
 #include "catalog/pg_type.h"
 #include "pg_lake/extensions/pg_lake_engine.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 #include "parser/parse_func.h"
 #include "utils/builtins.h"
 #include "utils/lsyscache.h"
