+++ pg_lake_engine/src/extensions/pg_lake_iceberg.c
@@ -26,7 +26,7 @@
 #include "utils/lsyscache.h"
 
 #include "pg_lake/extensions/pg_lake_iceberg.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 
 
 /*
