+++ pg_lake_engine/src/init.c
@@ -39,7 +39,8 @@
 #include "pg_lake/extensions/pg_lake_replication.h"
 #include "pg_lake/extensions/pg_parquet.h"
 #include "pg_lake/extensions/postgis.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
+#include "pg_extension_base/pg_extension_base_ids.h"
 #include "pg_lake/pgduck/cache_worker.h"
 #include "pg_lake/pgduck/client.h"
 #include "utils/guc.h"
@@ -173,7 +174,7 @@ _PG_init(void)
 		InitializePgMapIdCache();
 		InitializePgLakeSpatialIdCache();
 		InitializePgLakeBenchmarkIdCache();
-		InitializePGExtensionBaseCache();
+		InitializePgExtensionBaseCache();
 		InitializePgLakeIdCache();
 		InitializePgLakeTableIdCache();
 		InitializePgLakeIcebergIdCache();
