+++ pg_lake_table/src/ddl/drop_table.c
@@ -35,7 +35,7 @@
 #include "pg_lake/ddl/ddl_changes.h"
 #include "pg_lake/ddl/drop_table.h"
 #include "pg_lake/ddl/utility_hook.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 #include "pg_lake/extensions/pg_lake_table.h"
 #include "pg_lake/extensions/pg_lake_iceberg.h"
 #include "pg_lake/cleanup/deletion_queue.h"
@@ -50,7 +50,7 @@
 #include "pg_lake/iceberg/metadata_operations.h"
 #include "pg_lake/iceberg/operations/find_referenced_files.h"
 #include "pg_lake/util/rel_utils.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 #include "pg_lake/object_store_catalog/object_store_catalog.h"
 #include "pg_lake/transaction/track_iceberg_metadata_changes.h"
 #include "pg_lake/rest_catalog/rest_catalog.h"
