+++ pg_lake_table/src/ddl/vacuum.c
@@ -53,7 +53,7 @@
 #include "pg_lake/transaction/track_iceberg_metadata_changes.h"
 #include "pg_lake/object_store_catalog/object_store_catalog.h"
 #include "pg_lake/rest_catalog/rest_catalog.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 #include "nodes/makefuncs.h"
 #include "nodes/pg_list.h"
 #include "utils/acl.h"
