+++ pg_lake_table/src/fdw/data_file_stats_catalog.c
@@ -25,7 +25,7 @@
 #include "pg_lake/extensions/pg_lake_table.h"
 #include "pg_lake/fdw/data_files_catalog.h"
 #include "pg_lake/fdw/data_file_stats_catalog.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 
 #include "catalog/namespace.h"
 #include "utils/lsyscache.h"
