+++ pg_lake_table/src/fdw/data_files_catalog.c
@@ -29,7 +29,7 @@
 #include "pg_lake/data_file/data_files.h"
 #include "pg_lake/data_file/data_file_stats.h"
 #include "pg_lake/extensions/pg_lake_table.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 #include "pg_lake/extensions/pg_lake_engine.h"
 #include "pg_lake/fdw/catalog/row_id_mappings.h"
 #include "pg_lake/fdw/data_files_catalog.h"
@@ -55,7 +55,7 @@
 #include "pg_lake/util/array_utils.h"
 #include "pg_lake/util/plan_cache.h"
 #include "pg_lake/util/s3_reader_utils.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 #include "pg_lake/util/string_utils.h"
 #include "executor/spi.h"
 #include "foreign/foreign.h"
