+++ pg_lake_table/src/fdw/table_size.c
@@ -22,7 +22,7 @@
 #include "fmgr.h"
 
 #include "pg_lake/extensions/pg_lake_table.h"
-#include "pg_lake/extensions/extension_ids.h"
+#include "pg_extension_base/extension_ids.h"
 #include "pg_lake/fdw/data_files_catalog.h"
 #include "pg_lake/util/rel_utils.h"
 #include "utils/builtins.h"
