+++ pg_lake_table/src/fdw/writable_table.c
@@ -54,7 +54,7 @@
 #include "pg_lake/pgduck/write_data.h"
 #include "pg_lake/transaction/track_iceberg_metadata_changes.h"
 #include "pg_lake/util/rel_utils.h"
-#include "pg_lake/util/spi_helpers.h"
+#include "pg_extension_base/spi_helpers.h"
 #include "pg_lake/util/string_utils.h"
 #include "foreign/foreign.h"
 #include "nodes/makefuncs.h"
